src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile src/requirements.in
#
alembic==1.8.1
    # via mlflow
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   databricks-cli
    #   flask
    #   mlflow
cloudpickle==2.1.0
    # via mlflow
coloredlogs==15.0.1
    # via onnxruntime
cycler==0.11.0
    # via matplotlib
databricks-cli==0.17.0
    # via mlflow
docker==5.0.3
    # via mlflow
entrypoints==0.4
    # via mlflow
flask==2.1.3
    # via
    #   mlflow
    #   prometheus-flask-exporter
flatbuffers==2.0
    # via onnxruntime
fonttools==4.34.4
    # via matplotlib
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via mlflow
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via mlflow
humanfriendly==10.0
    # via coloredlogs
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via mlflow
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kiwisolver==1.4.4
    # via matplotlib
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.5.2
    # via -r src/requirements.in
mlflow==1.27.0
    # via -r src/requirements.in
mpmath==1.2.1
    # via sympy
numpy==1.23.1
    # via
    #   matplotlib
    #   mlflow
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   scipy
oauthlib==3.2.0
    # via databricks-cli
onnx==1.12.0
    # via -r src/requirements.in
onnxruntime==1.12.0
    # via -r src/requirements.in
opencv-python==4.6.0.66
    # via -r src/requirements.in
packaging==21.3
    # via
    #   matplotlib
    #   mlflow
    #   onnxruntime
pandas==1.4.3
    # via mlflow
pillow==9.2.0
    # via matplotlib
prometheus-client==0.14.1
    # via prometheus-flask-exporter
prometheus-flask-exporter==0.20.2
    # via mlflow
protobuf==3.20.1
    # via
    #   mlflow
    #   onnx
    #   onnxruntime
pyjwt==2.4.0
    # via databricks-cli
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.1
    # via
    #   mlflow
    #   pandas
pyyaml==6.0
    # via mlflow
querystring-parser==1.2.4
    # via mlflow
requests==2.28.1
    # via
    #   databricks-cli
    #   docker
    #   mlflow
scipy==1.9.0
    # via mlflow
six==1.16.0
    # via
    #   databricks-cli
    #   python-dateutil
    #   querystring-parser
smmap==5.0.0
    # via gitdb
sqlalchemy==1.4.39
    # via
    #   alembic
    #   mlflow
sqlparse==0.4.2
    # via mlflow
sympy==1.10.1
    # via onnxruntime
tabulate==0.8.10
    # via databricks-cli
toml==0.10.2
    # via -r src/requirements.in
torch==1.12.0
    # via -r src/requirements.in
typing-extensions==4.3.0
    # via
    #   onnx
    #   torch
urllib3==1.26.11
    # via requests
websocket-client==1.3.3
    # via docker
werkzeug==2.2.1
    # via flask
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
